+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - Added reflection argument to format callbacks on `rebuildSearch` options.
+- Added new navigation and search functions that play nice when used by multiple plugins. Deprecated old navigation and search functions.
 
 ## V1.2.0